requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
defusedxml==0.7.1
Django==5.1.6
djangorestframework==3.15.2
djoser==2.3.1
idna==3.10
oauthlib==3.2.2
pycparser==2.22
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
psycopg2-binary==2.9.10

